requirements.txt
pandas
numpy
regex
emoji
matplotlib
wordcloud
streamlit
python-dateutil
dateparser
textblob    